requirements.txt
appnope==0.1.0
decorator==4.0.10
dj-database-url==0.4.1
Django==1.10.2
django-bootstrap3==7.1.0
django-registration==2.1.2
djangorestframework==3.4.7
factory-boy==2.7.0
fake-factory==0.7.2
flake8==3.0.4
httpie==0.9.6
ipython==5.1.0
ipython-genutils==0.1.0
mccabe==0.5.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.1
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
pycodestyle==2.0.0
pyflakes==1.2.3
Pygments==2.1.3
python-dateutil==2.5.3
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
